string_util: Resolve implicit signed-unsigned conversion
[Why]
If compiled with clang, which complains about:
> include/gul14/string_util.h:154:27: warning: implicit conversion changes signedness: 'typename iterator_traits<unsigned char *>::difference_type' (aka 'long') to 'const std::size_t' (aka 'const unsigned long') [-Wsign-conversion]
>   154 |     const std::size_t n = std::distance(begin, end);
>       |                       ~   ^~~~~~~~~~~~~~~~~~~~~~~~~

+++ include/gul14/string_util.h
@@ -151,7 +151,7 @@ template <typename Iterator>
 inline std::string
 hex_string(Iterator begin, Iterator end, string_view separator = "")
 {
-    const std::size_t n = std::distance(begin, end);
+    const std::size_t n = static_cast<std::size_t>(std::distance(begin, end));
 
     std::string result;
